pyproject.toml
[project]
name = "spire-pdf-mcp-server"
version = "0.1.1"
description = "MCP server for Pdf document manipulation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "Spire.Pdf.Free>=10.12.0"
]
[[project.authors]]
name = "e-iceblue"
email = "sales@e-iceblue.com"

[project.urls]
Homepage = "https://www.e-iceblue.com/"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
spire-pdf-mcp-server = "spire_pdf_mcp.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/spire_pdf_mcp"]

[tool.hatch.build]
packages = ["src/spire_pdf_mcp"]